requirements.txt
Django==2.2.17
django-cors-headers==3.5.0
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
